update function move success
